refactor(step4): input, output 패키지 추가
