+++ src/main/java/step3/Step3Application.java
@@ -1,8 +1,8 @@
 package step3;
 
 import step3.game.GameController;
-import step3.view.ConsoleInputView;
-import step3.view.ConsoleOutputView;
+import step3.view.input.ConsoleInputView;
+import step3.view.output.ConsoleOutputView;
 
 public class Step3Application {
     public static void main(String[] args) {
