+++ src/main/java/step3/view/input/ConsoleInputView.java
@@ -1,4 +1,4 @@
-package step3.view;
+package step3.view.input;
 
 import java.util.Scanner;
